fix fatal error

+++ upgrade.php
@@ -45,9 +45,7 @@ function rsssl_upgrade() {
 		}
 	}
 
-	if (
-		( $prev_version && version_compare( $prev_version, '6.0.0', '<' ) ) )
-	) {
+	if ( $prev_version && version_compare( $prev_version, '6.0.0', '<' ) ) {
 		delete_option('rsssl_admin_notices');
 		update_option('rsssl_show_onboarding', true, false);
 		//upgrade both site and network settings
